requirements.txt
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.0.0